+++ src/main/kotlin/me/bossm0n5t3r/leetcode/pascalsTriangle/PascalsTriangle.kt
@@ -1,0 +1,23 @@
+package me.bossm0n5t3r.leetcode.pascalsTriangle
+
+class PascalsTriangle {
+    class Solution {
+        fun generate(numRows: Int): List<List<Int>> {
+            if (numRows == 1) return listOf(listOf(1))
+            if (numRows == 2) return listOf(listOf(1), listOf(1, 1))
+            val result = mutableListOf(listOf(1), listOf(1, 1))
+            for (i in 3..numRows) {
+                val before = result.last()
+                val next = mutableListOf(1)
+                for (index in before.indices) {
+                    when (index) {
+                        before.size - 1 -> next.add(before[index])
+                        else -> next.add(before[index] + before[index + 1])
+                    }
+                }
+                result.add(next)
+            }
+            return result
+        }
+    }
+}
